+++ src/pymodule/py_pingtools/c_beamselection.cpp
@@ -27,8 +27,8 @@ void init_c_beamselection(pybind11::module& m)
         m, "BeamSelection", DOC(themachinethatgoesping, echosounders, pingtools, BeamSelection))
         .def(py::init<>(), DOC_BeamSelection(BeamSelection))
         .def(py::init<const BeamSampleSelection& >(), DOC_BeamSelection(BeamSelection), py::arg("beam_sample_selection"))
-        .def(py::init<uint16_t>(), DOC_BeamSelection(BeamSelection_2), py::arg("number_of_beams"))
-        .def(py::init<std::vector<uint16_t>>(),
+        .def(py::init<uint32_t>(), DOC_BeamSelection(BeamSelection_2), py::arg("number_of_beams"))
+        .def(py::init<std::vector<uint32_t>>(),
              DOC_BeamSelection(BeamSelection_3),
              py::arg("beam_numbers"))
